Added second machine: media_proxy
The media proxy bootstrap requires an argument(can be inserted in the config.json), its IP address, since it has to be specified for RTPProxy to work properly.

The media proxy machine relays RTP traffic, using RTPProxy as backend. The SDP in INVITEs etc. is rewritten to point to the RTPProxy instance.
There is no interaction between the balancer and the media_proxy yet, nor NAT is supported.